refactor: sns.js 파일을 components 디렉토리에서 constants 디렉토리로 이동
